# api/routers/admin_index_security.py
"""
MemoryOS Control Plane Router
=============================

This router handles all administrative, indexing, and security control endpoints.
Control plane requires elevated privileges and provides:
- System administration (users, spaces, policies)
- Index management (rebuild, optimization, status)
- Security operations (auditing, compliance, monitoring)
- System metrics and operational controls

Security: Requires administrative authentication with elevated privileges
"""

from datetime import datetime, timedelta
from enum import Enum
from typing import Any, Dict, List, Optional
from uuid import uuid4

from fastapi import APIRouter, Query

# Direct imports for response schemas
from api.schemas.responses import HealthResponse, JobAccepted, StandardResponse

# Initialize router with control plane configuration
router = APIRouter(
    prefix="/admin/v1",
    tags=["control"],
    responses={
        401: {"description": "Authentication required"},
        403: {"description": "Administrative privileges required"},
        404: {"description": "Resource not found"},
        429: {"description": "Rate limit exceeded"},
        500: {"description": "Internal server error"},
    },
)


class IndexType(str, Enum):
    """Index types for management operations."""

    SEMANTIC = "semantic"
    TEMPORAL = "temporal"
    SPATIAL = "spatial"
    GRAPH = "graph"
    FULL_TEXT = "full_text"


class SecurityLevel(str, Enum):
    """Security audit levels."""

    LOW = "low"
    MEDIUM = "medium"
    HIGH = "high"
    CRITICAL = "critical"


# TODO: Add admin authentication dependency
# async def get_admin_user(token: str = Depends(admin_auth)) -> UserProfile:
#     """Extract admin user with elevated privileges"""
#     pass

# ============================================================================
# SYSTEM ADMINISTRATION - User and space management
# ============================================================================


@router.get("/users")
async def list_users(
    limit: int = Query(50, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    active_only: bool = Query(True),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    List all system users.

    Administrative endpoint for user management,
    monitoring, and system oversight.
    """
    # TODO: Query user management service
    mock_users = [
        {
            "user_id": f"user_{i:03d}",
            "email": f"user{i}@example.com",
            "name": f"User {i}",
            "status": "active" if i % 10 != 0 else "inactive",
            "created_at": (datetime.now() - timedelta(days=i * 10)).isoformat(),
            "last_login": (datetime.now() - timedelta(hours=i)).isoformat(),
            "space_count": 2 + (i % 5),
            "memory_count": 100 + (i * 50),
            "security_level": "standard",
        }
        for i in range(1, min(limit + 1, 21))  # Max 20 mock users
    ]

    if active_only:
        mock_users = [u for u in mock_users if u["status"] == "active"]

    return {
        "users": mock_users,
        "total_count": len(mock_users),
        "active_count": len([u for u in mock_users if u["status"] == "active"]),
        "pagination": {
            "limit": limit,
            "offset": offset,
            "has_more": len(mock_users) >= limit,
        },
    }


@router.get("/users/{user_id}")
async def get_user_details(
    user_id: str,
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get detailed user information.

    Administrative view of user data including
    security context, activity, and system usage.
    """
    # TODO: Query user service with admin context
    return {
        "user_id": user_id,
        "profile": {
            "email": f"{user_id}@example.com",
            "name": f"User {user_id}",
            "status": "active",
            "created_at": datetime.now().isoformat(),
            "last_login": datetime.now().isoformat(),
        },
        "security": {
            "security_level": "standard",
            "failed_logins": 0,
            "mfa_enabled": True,
            "last_security_review": datetime.now().isoformat(),
        },
        "usage": {
            "memory_spaces": 3,
            "total_memories": 1250,
            "storage_used_mb": 45.2,
            "api_calls_24h": 156,
            "last_activity": datetime.now().isoformat(),
        },
        "permissions": {
            "spaces": ["personal:default", "personal:work", "shared:team"],
            "capabilities": ["READ", "WRITE", "SHARE"],
            "restrictions": [],
        },
    }


@router.put("/users/{user_id}/status")
async def update_user_status(
    user_id: str,
    status_data: Dict[str, Any],
    # admin_user: UserProfile = Depends(get_admin_user)
) -> StandardResponse:
    """
    Update user status (activate, deactivate, suspend).

    Administrative control for user account management
    with audit logging and notification.
    """
    new_status = status_data.get("status", "active")
    reason = status_data.get("reason", "Administrative action")

    # TODO: Update user status in user service
    # TODO: Log admin action for audit
    # TODO: Send user notification if required

    return StandardResponse(
        success=True,
        message=f"User {user_id} status updated to {new_status}",
        timestamp=datetime.now(),
        data={
            "user_id": user_id,
            "previous_status": "active",
            "new_status": new_status,
            "reason": reason,
            "updated_by": "admin_user_id",
        },
    )


@router.get("/spaces")
async def list_memory_spaces(
    limit: int = Query(100, ge=1, le=1000),
    space_type: Optional[str] = Query(None),
    security_band: Optional[str] = Query(None),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    List all memory spaces in the system.

    Administrative view of space allocation, usage,
    and security configuration across the system.
    """
    # TODO: Query space registry with admin context
    mock_spaces = [
        {
            "space_id": f"personal:user_{i:03d}_default",
            "owner": f"user_{i:03d}",
            "type": "personal",
            "security_band": "GREEN",
            "created_at": (datetime.now() - timedelta(days=i * 5)).isoformat(),
            "memory_count": 50 + (i * 10),
            "storage_mb": 12.5 + (i * 2.3),
            "last_activity": (datetime.now() - timedelta(hours=i)).isoformat(),
            "status": "active",
        }
        for i in range(1, min(limit + 1, 11))
    ]

    # Add some shared spaces
    mock_spaces.extend(
        [
            {
                "space_id": "shared:team_alpha",
                "owner": "system",
                "type": "shared",
                "security_band": "AMBER",
                "created_at": (datetime.now() - timedelta(days=30)).isoformat(),
                "memory_count": 340,
                "storage_mb": 78.9,
                "last_activity": datetime.now().isoformat(),
                "status": "active",
                "members": 8,
            },
            {
                "space_id": "system:global_index",
                "owner": "system",
                "type": "system",
                "security_band": "RED",
                "created_at": (datetime.now() - timedelta(days=365)).isoformat(),
                "memory_count": 15000,
                "storage_mb": 2345.6,
                "last_activity": datetime.now().isoformat(),
                "status": "active",
            },
        ]
    )

    # Apply filters
    if space_type:
        mock_spaces = [s for s in mock_spaces if s["type"] == space_type]
    if security_band:
        mock_spaces = [s for s in mock_spaces if s["security_band"] == security_band]

    return {
        "spaces": mock_spaces,
        "total_count": len(mock_spaces),
        "summary": {
            "by_type": {
                "personal": len([s for s in mock_spaces if s["type"] == "personal"]),
                "shared": len([s for s in mock_spaces if s["type"] == "shared"]),
                "system": len([s for s in mock_spaces if s["type"] == "system"]),
            },
            "by_security_band": {
                "GREEN": len([s for s in mock_spaces if s["security_band"] == "GREEN"]),
                "AMBER": len([s for s in mock_spaces if s["security_band"] == "AMBER"]),
                "RED": len([s for s in mock_spaces if s["security_band"] == "RED"]),
            },
        },
    }


# ============================================================================
# INDEX MANAGEMENT - Search and retrieval optimization
# ============================================================================


@router.get("/indices/status")
async def get_index_status(
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get status of all system indices.

    Administrative view of index health, performance,
    and optimization status across all index types.
    """
    # TODO: Query index management service
    return {
        "indices": {
            "semantic": {
                "status": "healthy",
                "size_mb": 1250.5,
                "documents": 25000,
                "last_updated": datetime.now().isoformat(),
                "optimization_level": 95,
                "query_performance_ms": 45,
            },
            "temporal": {
                "status": "healthy",
                "size_mb": 890.2,
                "documents": 25000,
                "last_updated": datetime.now().isoformat(),
                "optimization_level": 88,
                "query_performance_ms": 12,
            },
            "spatial": {
                "status": "degraded",
                "size_mb": 567.8,
                "documents": 18500,
                "last_updated": (datetime.now() - timedelta(hours=6)).isoformat(),
                "optimization_level": 65,
                "query_performance_ms": 120,
                "issues": ["fragmentation detected", "requires optimization"],
            },
            "graph": {
                "status": "healthy",
                "size_mb": 2340.1,
                "documents": 25000,
                "relationships": 125000,
                "last_updated": datetime.now().isoformat(),
                "optimization_level": 92,
                "query_performance_ms": 89,
            },
            "full_text": {
                "status": "healthy",
                "size_mb": 3450.7,
                "documents": 25000,
                "last_updated": datetime.now().isoformat(),
                "optimization_level": 98,
                "query_performance_ms": 23,
            },
        },
        "overall_health": "good",
        "total_size_mb": 8498.3,
        "recommendations": [
            "Schedule optimization for spatial index",
            "Consider increasing graph index cache size",
        ],
    }


@router.post("/indices/{index_type}/rebuild", response_model=JobAccepted)
async def rebuild_index(
    index_type: IndexType,
    force: bool = Query(False, description="Force rebuild even if healthy"),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> JobAccepted:
    """
    Rebuild a specific index.

    Administrative operation to rebuild indices for
    performance optimization or corruption recovery.
    """
    job_id = f"rebuild_{index_type.value}_{uuid4().hex[:8]}"

    # TODO: Route to index management service
    # TODO: Estimate rebuild time based on data size
    estimated_duration = {
        "semantic": 3600,  # 1 hour
        "temporal": 1800,  # 30 minutes
        "spatial": 2400,  # 40 minutes
        "graph": 5400,  # 1.5 hours
        "full_text": 7200,  # 2 hours
    }.get(index_type.value, 3600)

    return JobAccepted(
        job_id=job_id,
        status="accepted",
        estimated_duration_seconds=estimated_duration,
        status_url=f"/admin/v1/jobs/{job_id}/status",
    )


@router.post("/indices/optimize", response_model=JobAccepted)
async def optimize_all_indices(
    # admin_user: UserProfile = Depends(get_admin_user)
) -> JobAccepted:
    """
    Optimize all system indices.

    Comprehensive optimization operation for
    system-wide performance improvement.
    """
    job_id = f"optimize_all_{uuid4().hex[:8]}"

    # TODO: Route to index management service
    return JobAccepted(
        job_id=job_id,
        status="accepted",
        estimated_duration_seconds=10800,  # 3 hours
        status_url=f"/admin/v1/jobs/{job_id}/status",
    )


# ============================================================================
# SECURITY OPERATIONS - Auditing and compliance
# ============================================================================


@router.get("/security/audit-log")
async def get_audit_log(
    limit: int = Query(100, ge=1, le=1000),
    level: Optional[SecurityLevel] = Query(None),
    start_date: Optional[str] = Query(None),
    end_date: Optional[str] = Query(None),
    user_id: Optional[str] = Query(None),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get security audit log entries.

    Administrative access to security events,
    compliance monitoring, and incident tracking.
    """
    # TODO: Query security audit service
    mock_events = [
        {
            "event_id": f"audit_{i:04d}",
            "timestamp": (datetime.now() - timedelta(hours=i)).isoformat(),
            "level": ["LOW", "MEDIUM", "HIGH", "CRITICAL"][i % 4],
            "event_type": ["LOGIN", "ACCESS_DENIED", "POLICY_VIOLATION", "DATA_ACCESS"][
                i % 4
            ],
            "user_id": f"user_{(i % 10) + 1:03d}",
            "resource": f"space:personal:user_{(i % 10) + 1:03d}",
            "action": "READ",
            "result": "success" if i % 5 != 0 else "failure",
            "ip_address": f"192.168.1.{(i % 254) + 1}",
            "user_agent": "MemoryOS-App/1.0.0",
            "details": f"Security event {i} details",
        }
        for i in range(min(limit, 50))
    ]

    # Apply filters
    if level:
        mock_events = [e for e in mock_events if e["level"] == level.value.upper()]
    if user_id:
        mock_events = [e for e in mock_events if e["user_id"] == user_id]

    return {
        "events": mock_events,
        "total_count": len(mock_events),
        "summary": {
            "by_level": {
                "critical": len([e for e in mock_events if e["level"] == "CRITICAL"]),
                "high": len([e for e in mock_events if e["level"] == "HIGH"]),
                "medium": len([e for e in mock_events if e["level"] == "MEDIUM"]),
                "low": len([e for e in mock_events if e["level"] == "LOW"]),
            },
            "by_result": {
                "success": len([e for e in mock_events if e["result"] == "success"]),
                "failure": len([e for e in mock_events if e["result"] == "failure"]),
            },
        },
    }


@router.get("/security/compliance-report")
async def get_compliance_report(
    report_type: str = Query("summary", regex="^(summary|detailed|export)$"),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get security compliance report.

    Administrative reporting for regulatory compliance,
    security posture assessment, and audit preparation.
    """
    # TODO: Query compliance service
    return {
        "report_type": report_type,
        "generated_at": datetime.now().isoformat(),
        "period": {
            "start": (datetime.now() - timedelta(days=30)).isoformat(),
            "end": datetime.now().isoformat(),
        },
        "compliance_score": 94.5,
        "findings": {"passed": 127, "warnings": 8, "violations": 2},
        "categories": {
            "data_protection": {"score": 96, "status": "compliant"},
            "access_control": {"score": 98, "status": "compliant"},
            "audit_logging": {"score": 89, "status": "warning"},
            "encryption": {"score": 100, "status": "compliant"},
            "retention_policy": {"score": 85, "status": "warning"},
        },
        "recommendations": [
            "Review audit log retention policy",
            "Update access control documentation",
            "Schedule quarterly security review",
        ],
        "violations": [
            {
                "id": "V001",
                "severity": "medium",
                "description": "Audit log gap detected",
                "affected_systems": ["index_service"],
                "remediation": "Verify audit configuration",
            }
        ],
    }


@router.post("/security/scan", response_model=JobAccepted)
async def initiate_security_scan(
    scan_type: str = Query("comprehensive", regex="^(quick|comprehensive|targeted)$"),
    target_resources: Optional[List[str]] = Query(None),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> JobAccepted:
    """
    Initiate security vulnerability scan.

    Administrative security assessment across
    system components, configurations, and data.
    """
    job_id = f"security_scan_{scan_type}_{uuid4().hex[:8]}"

    # TODO: Route to security scanning service
    estimated_duration = {
        "quick": 900,  # 15 minutes
        "comprehensive": 7200,  # 2 hours
        "targeted": 1800,  # 30 minutes
    }.get(scan_type, 3600)

    return JobAccepted(
        job_id=job_id,
        status="accepted",
        estimated_duration_seconds=estimated_duration,
        status_url=f"/admin/v1/jobs/{job_id}/status",
    )


# ============================================================================
# SYSTEM METRICS - Operational monitoring
# ============================================================================


@router.get("/metrics/system")
async def get_system_metrics(
    timeframe: str = Query("1h", regex="^(5m|15m|1h|6h|24h|7d)$"),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get system operational metrics.

    Administrative monitoring of system performance,
    resource utilization, and operational health.
    """
    # TODO: Query metrics service
    return {
        "timeframe": timeframe,
        "collected_at": datetime.now().isoformat(),
        "system": {
            "cpu_usage_percent": 45.2,
            "memory_usage_percent": 67.8,
            "disk_usage_percent": 34.1,
            "network_io_mbps": 12.5,
            "active_connections": 1250,
        },
        "services": {
            "memory_store": {
                "status": "healthy",
                "response_time_ms": 45,
                "error_rate": 0.1,
            },
            "query_engine": {
                "status": "healthy",
                "response_time_ms": 89,
                "error_rate": 0.2,
            },
            "index_service": {
                "status": "healthy",
                "response_time_ms": 23,
                "error_rate": 0.0,
            },
            "auth_service": {
                "status": "healthy",
                "response_time_ms": 12,
                "error_rate": 0.1,
            },
        },
        "api": {
            "requests_per_minute": 1567,
            "average_response_time_ms": 156,
            "error_rate_percent": 0.8,
            "p95_response_time_ms": 234,
            "active_sessions": 89,
        },
        "storage": {
            "total_memories": 25000,
            "storage_used_gb": 12.5,
            "daily_growth_mb": 45.2,
            "backup_status": "healthy",
            "replication_lag_ms": 23,
        },
    }


@router.get("/metrics/usage")
async def get_usage_metrics(
    period: str = Query("24h", regex="^(1h|24h|7d|30d)$"),
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get system usage analytics.

    Administrative insights into user activity,
    feature utilization, and growth patterns.
    """
    # TODO: Query analytics service
    return {
        "period": period,
        "users": {
            "total_registered": 1250,
            "active_users": 890,
            "new_registrations": 23,
            "retention_rate": 85.5,
        },
        "memories": {
            "total_created": 25000,
            "new_memories": 450,
            "average_per_user": 28.1,
            "popular_types": ["text", "document", "image", "audio"],
        },
        "api_usage": {
            "total_requests": 156789,
            "by_endpoint": {
                "/api/v1/memories/search": 45234,
                "/api/v1/memories": 23456,
                "/v1/tools/memory/recall": 34567,
                "/v1/tools/memory/submit": 12345,
            },
            "by_plane": {"app": 65.2, "agents": 28.5, "control": 6.3},
        },
        "growth_trends": {
            "user_growth_rate": 12.5,
            "memory_growth_rate": 34.2,
            "storage_growth_rate": 15.8,
        },
    }


# ============================================================================
# JOB MANAGEMENT - Background operations
# ============================================================================


@router.get("/jobs/{job_id}/status")
async def get_job_status(
    job_id: str,
    # admin_user: UserProfile = Depends(get_admin_user)
) -> Dict[str, Any]:
    """
    Get status of background job.

    Administrative monitoring of long-running operations
    like index rebuilds, security scans, and data migrations.
    """
    # TODO: Query job management service
    return {
        "job_id": job_id,
        "status": "running",
        "progress_percent": 65,
        "started_at": (datetime.now() - timedelta(minutes=45)).isoformat(),
        "estimated_completion": (datetime.now() + timedelta(minutes=25)).isoformat(),
        "operation": "index_rebuild",
        "details": {
            "index_type": "semantic",
            "documents_processed": 16250,
            "total_documents": 25000,
            "current_phase": "optimization",
        },
        "logs": [
            "Starting semantic index rebuild",
            "Processing documents: 16250/25000",
            "Optimization phase initiated",
            "ETA: 25 minutes remaining",
        ],
    }


# ============================================================================
# HEALTH AND STATUS
# ============================================================================


@router.get("/health", response_model=HealthResponse)
async def control_plane_health() -> HealthResponse:
    """
    Health check endpoint for control plane.

    Returns operational status of administrative services
    and control plane components.
    """
    return HealthResponse(
        status="healthy",
        timestamp=datetime.now(),
        version="1.0.0",
        uptime_seconds=172800,  # 48 hours
        dependencies={
            "user_management": "healthy",
            "space_registry": "healthy",
            "index_manager": "healthy",
            "security_service": "healthy",
            "audit_logger": "healthy",
            "metrics_collector": "healthy",
        },
    )