# contracts/api/app/users.openapi.v1.yaml
openapi: 3.1.0
info:
  title: MemoryOS App Plane - Users API
  description: |
    **Memory-Centric Family AI User Management**

    This API manages the complete user lifecycle in the Memory-Centric Family AI system, including:
    - **Family Structure Management**: Family units, members, relationships, and hierarchies
    - **Device Ecosystem**: User devices with memory capabilities and sync coordination
    - **Memory Spaces**: Personal, selective, shared, extended, and interfamily memory organization
    - **User Presence**: Real-time presence and family coordination
    - **Profile Management**: Family-aware user profiles with relationship contexts

    **Memory-Centric Design**: All endpoints integrate with the Memory Backbone to ensure:
    - Family relationships are preserved in memory formation
    - User context flows through memory operations
    - Device capabilities inform memory sync strategies
    - Family intelligence guides user experience personalization

    **Architecture Integration**:
    - Serves the Memory Backbone as the central nervous system
    - Integrates with Family Intelligence for relationship awareness
    - Coordinates with Device Ecosystem for cross-device memory sync
    - Supports 3 API Planes: App, Agent, and Control serving memory operations
  version: "1.0.0"
  contact:
    name: MemoryOS API Team
    url: https://familyos.local/support
    email: api-support@familyos.local
  license:
    name: Family OS Private License
...
servers:
...
security:
  - FamilyOSAuth: []
  - DeviceAuth: []
paths:
  /users:
    get:
      summary: List Family Members
      description: |
        Retrieve family members with relationship context and memory access permissions.

        **Memory Integration**: Returns users with their memory space access and family relationship context.
        **Family Intelligence**: Respects family hierarchy and relationship-based visibility.
      operationId: listFamilyMembers
      parameters:
        - name: family_id
          in: query
          description: Family unit identifier for filtering
          required: false
          schema:
            type: string
            format: uuid
        - name: relationship_context
          in: query
          description: Relationship context filter (e.g., 'ab', 'abc', 'family')
          required: false
          schema:
            type: string
            pattern: '^[a-z]+$'
        - name: include_memory_spaces
          in: query
          description: Include user memory space access information
          required: false
          schema:
            type: boolean
            default: false
        - name: presence_filter
          in: query
          description: Filter by presence status
          required: false
          schema:
            type: string
            enum: [online, offline, away, busy]
      responses:
        '200':
          description: Family members retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyUser'
                  family_context:
                    $ref: '#/components/schemas/FamilyContext'
                  memory_summary:
                    $ref: '#/components/schemas/MemorySummary'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Family Management

    post:
      summary: Add Family Member
      description: |
        Add a new member to the family with specified role and memory access permissions.

        **Memory Integration**: Creates user memory spaces and establishes family relationship context.
        **Family Intelligence**: Configures family hierarchy and relationship-based access control.
      operationId: addFamilyMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFamilyUserRequest'
      responses:
        '201':
          description: Family member added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/FamilyUser'
                  memory_spaces_created:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemorySpace'
                  family_relationships_established:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyRelationship'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: User already exists in family
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Family Management
  /users/{user_id}:
    get:
      summary: Get Family Member Details
      description: |
        Retrieve detailed information about a specific family member including relationship context and memory access.

        **Memory Integration**: Returns memory space access, recent memory activity, and family context.
        **Family Intelligence**: Provides relationship-aware information with appropriate privacy controls.
      operationId: getFamilyMember
      parameters:
        - name: user_id
          in: path
          description: Family member identifier
          required: true
          schema:
            type: string
            format: uuid
        - name: include_memory_activity
          in: query
          description: Include recent memory activity summary
          required: false
          schema:
            type: boolean
            default: false
        - name: include_devices
          in: query
          description: Include user device information
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Family member details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/FamilyUserDetailed'
                  relationship_context:
                    $ref: '#/components/schemas/RelationshipContext'
                  memory_activity:
                    $ref: '#/components/schemas/MemoryActivity'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDevice'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Family Management

    put:
      summary: Update Family Member
      description: |
        Update family member information including role, permissions, and memory access.

        **Memory Integration**: Updates memory space access and family relationship context.
        **Family Intelligence**: Maintains family relationship consistency and privacy controls.
      operationId: updateFamilyMember
      parameters:
        - name: user_id
          in: path
          description: Family member identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFamilyUserRequest'
      responses:
        '200':
          description: Family member updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/FamilyUser'
                  changes_applied:
                    type: array
                    items:
                      type: string
                  memory_access_updated:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Family Management

    delete:
      summary: Remove Family Member
      description: |
        Remove a family member and handle memory migration/cleanup according to family policies.

        **Memory Integration**: Manages memory cleanup, migration to family memory spaces, and relationship updates.
        **Family Intelligence**: Maintains family relationship integrity and memory continuity.
      operationId: removeFamilyMember
      parameters:
        - name: user_id
          in: path
          description: Family member identifier
          required: true
          schema:
            type: string
            format: uuid
        - name: memory_migration_strategy
          in: query
          description: Strategy for handling user memories
          required: false
          schema:
            type: string
            enum: [migrate_to_family, archive_personal, delete_all]
            default: migrate_to_family
      responses:
        '200':
          description: Family member removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed_user_id:
                    type: string
                    format: uuid
                  memory_migration_summary:
                    $ref: '#/components/schemas/MemoryMigrationSummary'
                  family_relationships_updated:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyRelationship'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Family Management
...
components:
  securitySchemes:
    FamilyOSAuth:
      type: oauth2
      description: FamilyOS OAuth2 authentication with family-aware scopes
      flows:
        authorizationCode:
          authorizationUrl: https://auth.familyos.local/oauth/authorize
          tokenUrl: https://auth.familyos.local/oauth/token
          scopes:
            family:read: Read family information and relationships
            family:write: Modify family structure and relationships
            memory:read: Access family memory spaces
            memory:write: Create and modify family memories
            device:manage: Manage family devices and sync
            presence:read: View family member presence
            presence:write: Update presence status

    DeviceAuth:
      type: apiKey
      description: Device-specific authentication for registered family devices
      name: X-Device-Token
      in: header
  schemas:
...
    FamilyContext:
      type: object
      properties:
        family_id:
          type: string
          format: uuid
        relationship_map:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        emotional_context:
          type: object
          properties:
            family_mood:
              type: string
              enum: [positive, neutral, negative, mixed, unknown]
        coordination_active:
          type: boolean
    MemorySummary:
      type: object
      properties:
        total_memories:
          type: integer
        memory_spaces:
          type: integer
        recent_activity:
          type: integer
        shared_memories:
          type: integer
...
  responses:
...
    ForbiddenError:
      description: Insufficient permissions or family access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...